+++ src/btron.rs
@@ -1,3 +1,4 @@
+pub mod allocator;
 mod brightv;
 pub mod rustify;
 pub mod types;
