+++ src/lib.rs
@@ -1,3 +1,4 @@
 #![no_std]
+#![feature(c_size_t)]
 
 pub mod btron;
